+++ src/utils/utilities.h
@@ -20,15 +20,15 @@
 
 class Body;
 
-inline long double DistSquared(long double x1, long double y1, long double x2, long double y2) {
+inline double DistSquared(double x1, double y1, double x2, double y2) {
     return std::pow(x2 - x1, 2) + std::pow(y2 - y1, 2);
 }
 
-inline long double DistSquared(const vec3 &p1, const vec3 &p2) {
+inline double DistSquared(const vec3 &p1, const vec3 &p2) {
     return std::pow(p1.x - p2.x, 2) + std::pow(p1.y - p2.y, 2) + std::pow(p1.z - p2.z, 2);
 }
 
-inline long double NewtonsGravitationalLaw(long double m1, long double m2, long double rSquared) {
+inline double NewtonsGravitationalLaw(double m1, double m2, double rSquared) {
     return G * m1 * m2 * (1 / (rSquared));
 }
 
@@ -38,4 +38,4 @@ inline long double NewtonsGravitationalLaw(long double m1, long double m2, long 
  * @param P the magitude of the positions
  * @return the maximum possible position
  */
-long double GetMaxPosition(int P);
+double GetMaxPosition(int P);
